test(e2e): fold sigterm handoff coverage into existing failover test
Per review, drop the separate failover spec and extend the existing
'should detect and recover when a primary deployment is deleted' test
instead, so failover coverage lives in one place. The existing test
gains, around its StatefulSet-deletion disruption:

- a password-protected default user (passwordSecret, as in #292) so
  valkey-cli commands run authenticated
- 50 keys seeded before the disruption and verified after recovery
- a continuous writer through the disruption with per-attempt acks and
  a 2s connection timeout (a stale MOVED redirect to the terminated
  primary's IP otherwise hangs a connect for the ~130s TCP SYN timeout)
- promotion of the shard's replica asserted within the 30s termination
  grace period
- verification that every write acknowledged during the disruption is
  readable afterwards
- detection of whether the shutdown-on-sigterm handoff engaged (from
  the promoted replica's log), with the writer-gap bound asserted on
  the handoff path; the path itself is reported rather than
  hard-asserted because valkey 9.0's replica selection on shutdown
  requires exact ack-offset equality and intermittently falls back to
  failure detection

Full e2e suite on Kind: 41 of 41 specs passed. On the handoff path the
writer recorded 7904 acknowledged writes with a 0.23s longest gap and
zero lost.
